Use long long because otherwise it's going to be screwy on 32bit machines

+++ dht/dhtcore/Janitor.c
@@ -126,8 +126,8 @@ static void blacklist(struct Janitor_pvt* j, uint64_t path)
             oldestTime = qp->timeAdded;
         }
     }
-    Log_debug(j->logger, "Replacing [%ld]ms old blacklist node because blacklist is full",
-        (long)(now - oldestTime));
+    Log_debug(j->logger, "Replacing [%lld]ms old blacklist node because blacklist is full",
+        (long long)(now - oldestTime));
     j->blacklist[oldestIndex].timeAdded = now;
     j->blacklist[oldestIndex].path = path;
 }
